include propertyDefaults for ObservableArray
