+++ src/templates/can-property-definitions/observable-default-propertydefaults-output.js
