+++ test/fixtures/version-6/can-property-definitions/observable-default-propertydefaults-output.html
@@ -57,3 +57,41 @@
     }
   }
 </script>
+
+<script>
+  import { ObservableArray } from "can";
+  import DeepObservable from "can-deep-observable";
+
+  class Bar extends ObservableArray {
+    static get props() {
+      return {
+        items: {
+          type: Array,
+          default: Array
+        },
+
+        name: 'string'
+      };
+    }
+
+    static get items() {
+      return Item;
+    }
+  };
+
+  class Foo extends ObservableArray {
+    static get props() {
+      return {
+        items: {
+          type: Array,
+          default: Array
+        },
+        name: 'string'
+      };
+    }
+
+    static get items() {
+      return DeepObservable;
+    }
+  }
+</script>
