+++ test/fixtures/version-6/can-property-definitions/observable-default-propertydefaults-output.js
