+++ test/test.js
@@ -122,4 +122,4 @@ require('../lib/transforms/version-6/can-property-definitions/connected-events-t
 require('../lib/transforms/version-6/can-property-definitions/connectedcallback-test.js');
 require('../lib/transforms/version-6/can-define-array/array-constructor-test.js');
 require('../lib/transforms/version-6/can-define-object/constructor-test.js');
-require('../lib/transforms/version-6/can-property-definitions/observable-object-default-propertydefaults-test.js');
+require('../lib/transforms/version-6/can-property-definitions/observable-default-propertydefaults-test.js');
